backend/requirements.txt
alembic==1.11.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.95.2
greenlet==3.2.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.15
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.22.0
